+++ example/lib/theme.dart
@@ -65,6 +65,13 @@ class AppTheme extends ChangeNotifier {
     _locale = locale;
     notifyListeners();
   }
+
+  VisualDensity _visualDensity = VisualDensity.standard;
+  VisualDensity get visualDensity => _visualDensity;
+  set visualDensity(final VisualDensity density) {
+    _visualDensity = density;
+    notifyListeners();
+  }
 }
 
 AccentColor get systemAccentColor {