+++ lib/src/controls/pickers/time_picker.dart
@@ -317,7 +317,9 @@ class TimePickerState extends State<TimePicker>
             child: AnimatedContainer(
               duration: theme.fastAnimationDuration,
               curve: theme.animationCurve,
-              constraints: const BoxConstraints(minHeight: kPickerHeight),
+              constraints: BoxConstraints(
+                minHeight: (kPickerHeight + theme.visualDensity.baseSizeAdjustment.dy).clamp(0.0, double.infinity),
+              ),
               decoration: kPickerDecorationBuilder(context, states),
               child: DefaultTextStyle.merge(
                 style: TextStyle(